update html script to type=modules, update project dir to include css,img,js dirs, add createCard, genericRequests,urls modules, update README to include next phase 2 requirements
