+++ README.md
@@ -1,6 +1,6 @@
 # Tasks
 
-## DAY 2 
+## DAY 2 - Phase 1
 Agenda 
 - Continue work with WishList App
 
@@ -11,7 +11,7 @@ Agenda
                 -[ ] FOLLOW-UP: why button submit event did not work originally
         - [x] research header in vs out of body
 
-- [ ] TODO: CSS
+- [x] DONE: CSS
         - [x] zero out default web styles
         - [x] add click/hover effects to buttons
         - [x] match as close as possible: header fonts size/family
@@ -27,25 +27,32 @@ Agenda
             - By: rm hard code, ensure DRY, add helper methods
             - Replace: use regex for conditional (ln 21)
 
-## Day 3 
+## Day 3 - Phase-2
+
 Agenda: 
+
 - Refactor 
   - carryover day 2 JS Todos 
   - modularize index.js 
     - [js.io modules](https://javascript.info/modules-intro)
-- Plan Unsplash API integration
-  - how to fetch API data   
-    - [js.io fetch](https://javascript.info/fetch)
-    - [js.io network requests](https://javascript.info/network)
+
+- PHASE 2 REQUIREMENTS: Plan Unsplash API integration
+    - remove photo url field from WishList form
+    - using only destination/location fields
+      - fetch RELAVENT image from Unslash API
+      - use that as the rendered cards image
 
 - [ ] TODO: Research
+  - [ ] READ Javascript.io  
+      - [js.io fetch](https://javascript.info/fetch)
+      - [js.io network requests](https://javascript.info/network)
   - [ ] READ Unsplash Documentation
-    - register for api access
-      - access/secret key guidlines
-    - base url
-      - [relavent endpoint resources to be queried](https://unsplash.com/documentation#search-photos)
+      - register for api access
+        - access/secret key guidlines
+      - base url
+        - [relavent endpoint resources to be queried](https://unsplash.com/documentation#search-photos)
 
 - [ ] TODO: JS 
   - [ ] REFACTOR: 
-    - rm hard code, ensure DRY, add helper methods
-    - Replace: use regex for conditional (ln 21)
+      - rm hard code, ensure DRY, add helper methods
+      - Replace: use regex for conditional (ln 21)
