+++ js/urls.js
@@ -1,0 +1,8 @@
+"use strict";
+
+// const baseUrl = "https://jsonplaceholder.typicode.com/photos"  // 5k photos
+// const baseUrl = "https://jsonplaceholder.typicode.com/photo/10" // 1 specific photo
+const placeholderBaseURL = "https://jsonplaceholder.typicode.com/albums/1/photos"    // 50 photos 
+
+const baseURL = placeholderBaseURL
+export{ baseURL }
